Added if check for file existence

+++ create_lammps_pairstyle_table.py
@@ -93,8 +93,12 @@ def pair_write_lammps(lmps_input_filename,lmps_pair_filename,lmps_executing_comm
     lmpout = open("./"+lmps_input_filename,'w')
     lmpout.write(lmps_comms)
     lmpout.close()
+    if os.path.exists(lmps_pair_filename):
+        os.system("rm "+lmps_pair_filename+" && "+lmps_executing_command+" "+lmps_input_filename+" > lmps_pair_write_from_create_lammps_pairstyle_table.log") # Let's\
+ use os.system() because reasons...                                                                                                                                   
+    else:
+        os.system(lmps_executing_command+" "+lmps_input_filename+" > lmps_pair_write_from_create_lammps_pairstyle_table.log")
     
-    os.system("rm "+lmps_pair_filename+" && "+lmps_executing_command+" "+lmps_input_filename+" > lmps_pair_write_from_create_lammps_pairstyle_table.log") # Let's use os.system() because reasons...
 
 
 def comparison(file_basename,lmps_pair_filename,pair_filename,xmin,xmax,deltax,y1min,y1max,y2min,y2max,plot=True):
